# use get_puzzle_and_solution_for_coin() from chia_rs