This bumps the `chia_rs` dependency to 0.1.9, which has an implementation of `get_puzzle_and_solution_for_coin()`. It also switches over `chia-blockchain` to use that function.

One noteworthy difference is that the rust implementation returns `SerializedProgram` for both the puzzle and the solution. This ends up being more efficient in most cases since we mostly just turn around and serializes these over the wallet protocol.

To quantify the speedup, the `tests/core/test_cost_calculation.py::test_get_puzzle_and_solution_for_coin_performance` benchmark has the improvement.

| benchmark | before | after | after / before |
| --- | --- | --- | --- |
| get_puzzle_and_solution_for_coin() | 8.67s | 2.86s | 32.99% |

The output from the benchmark can be found here: https://github.com/Chia-Network/chia-blockchain/runs/8163694887?check_suite_focus=true#step:10:35

The rust implementation can be found here: https://github.com/Chia-Network/chia_rs/blob/main/src/gen/get_puzzle_and_solution.rs#L48